requirements.txt
allure-pytest==2.8.15
allure-python-commons==2.8.15
Appium-Python-Client==0.52
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
enum34==1.1.10
et-xmlfile==1.0.1
jdcal==1.4.1
lxml==4.5.1
more-itertools==8.3.0
namedlist==1.7
numpy==1.18.4
packaging==20.3
pandas==1.0.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
pytest-html==2.1.1
pytest-metadata==1.9.0
pytest-sugar==0.9.3
python-dateutil==2.8.1
pytz==2020.1
selenium==3.141.0
six==1.14.0
termcolor==1.1.0
urllib3==1.25.9
wcwidth==0.1.9
xlrd==1.2.0
